requirements.txt
# SignPaper Telegram Bot - Dependencies
python-telegram-bot>=20.7
python-dotenv>=1.0.0
supabase>=2.0.0
PyMuPDF>=1.23.0
aiohttp>=3.9.0
groq>=0.4.0
